# Lelkibeke/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Lelkibeke QR Restaurant API",
        "description": "API Documentation for Lelkibeke QR Restaurant",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/newTable": {
            "post": {
                "tags": [
                    "Tables"
                ],
                "summary": "Create a new table",
                "operationId": "30e5fa8630c604e7f1f61574de734648",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "table_number",
                                    "qr_code_url",
                                    "is_available"
                                ],
                                "properties": {
                                    "table_number": {
                                        "type": "integer"
                                    },
                                    "qr_code_url": {
                                        "type": "string"
                                    },
                                    "is_available": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Table created successfully"
                    },
                    "400": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/modifyTable": {
            "post": {
                "tags": [
                    "Tables"
                ],
                "summary": "Modify an existing table",
                "operationId": "4234eeea67d3b88b54c6da59d2d4d210",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id",
                                    "table_number",
                                    "qr_code_url",
                                    "is_available"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    },
                                    "table_number": {
                                        "type": "integer"
                                    },
                                    "qr_code_url": {
                                        "type": "string"
                                    },
                                    "is_available": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Table modified successfully"
                    },
                    "400": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/deleteTable": {
            "post": {
                "tags": [
                    "Tables"
                ],
                "summary": "Delete a table",
                "operationId": "35125b2da4f80c1899d4f9e3f783913b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Table deleted successfully"
                    },
                    "400": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/setOccupancyStatus": {
            "post": {
                "tags": [
                    "Tables"
                ],
                "summary": "Set occupancy status of a table",
                "operationId": "df52ee9266a57926b4513ea21b6f87b4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id",
                                    "is_available"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    },
                                    "is_available": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Occupancy status updated successfully"
                    },
                    "400": {
                        "description": "Validation error"
                    }
                }
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "bearer": []
        }
    ]
}